+++ src/Config.cpp
@@ -6,6 +6,7 @@
 bool Config::m_ignore_vjoy_version_warning = false;
 bool Config::m_enable_vjoy = false;
 bool Config::m_output_ffb_to_vjoy = false;
+std::string Config::m_last_device_guid = "";
 
 std::vector<Preset> Config::presets;
 
@@ -392,6 +393,7 @@ void Config::Save(const FFBEngine& engine, const std::string& filename) {
         file << "ignore_vjoy_version_warning=" << m_ignore_vjoy_version_warning << "\n";
         file << "enable_vjoy=" << m_enable_vjoy << "\n";
         file << "output_ffb_to_vjoy=" << m_output_ffb_to_vjoy << "\n";
+        file << "last_device_guid=" << m_last_device_guid << "\n";
         file << "gain=" << engine.m_gain << "\n";
         file << "sop_smoothing_factor=" << engine.m_sop_smoothing_factor << "\n";
         file << "slip_angle_smoothing=" << engine.m_slip_angle_smoothing << "\n";
@@ -492,6 +494,7 @@ void Config::Load(FFBEngine& engine, const std::string& filename) {
                     if (key == "ignore_vjoy_version_warning") m_ignore_vjoy_version_warning = std::stoi(value);
                     else if (key == "enable_vjoy") m_enable_vjoy = std::stoi(value);
                     else if (key == "output_ffb_to_vjoy") m_output_ffb_to_vjoy = std::stoi(value);
+                    else if (key == "last_device_guid") m_last_device_guid = value;
                     else if (key == "gain") engine.m_gain = std::stof(value);
                     else if (key == "sop_smoothing_factor") engine.m_sop_smoothing_factor = std::stof(value);
                     else if (key == "sop_scale") engine.m_sop_scale = std::stof(value);